replace long long type by int64_t
Former-commit-id: 80e5e903f14bcab07f60508451d5a0a098b52b62
Former-commit-id: e5b2b7113191a7c04e7799aa3c16a5cb0852ddab
Former-commit-id: 6844793d2d6c3485fd4fcef4ba82be9f8bc723ec [formerly 3e8405d8589c6c27e71a69e07a391f272f1c4e78]
Former-commit-id: 56e3512c6dd044ffd0beb5169239e24500a98d16
Former-commit-id: 20de22fca82f0af8460a93c9845327b0538d1080
Former-commit-id: 09e394d05a7758a22671273c4f765ab72770125a

+++ src/include/LASzip/mydefs.hpp
@@ -2,11 +2,11 @@
 ===============================================================================
 
   FILE:  mydefs.hpp
-  
+
   CONTENTS:
 
     Basic data type definitions and operations to be robust across platforms.
- 
+
   PROGRAMMERS:
 
     martin.isenburg@rapidlasso.com  -  http://rapidlasso.com
@@ -21,17 +21,19 @@
 
     This software is distributed WITHOUT ANY WARRANTY and without even the
     implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-  
+
   CHANGE HISTORY:
-  
+
     28 October 2015 -- adding DLL bindings via 'COMPILE_AS_DLL' and 'USE_AS_DLL'
     10 January 2011 -- licensing change for LGPL release and libLAS integration
     13 July 2005 -- created after returning with many mosquito bites from OBX
-  
+
 ===============================================================================
 */
 #ifndef MYDEFS_HPP
 #define MYDEFS_HPP
+
+#include <stdint.h>
 
 #ifndef _WIN32
 #define LASLIB_DLL
@@ -55,12 +57,12 @@ typedef unsigned int       U32;
 typedef unsigned short     U16;
 typedef unsigned char      U8;
 
-#if defined(_WIN32) && ! defined (__MINGW32__) // 64 byte integer under Windows 
+#if defined(_WIN32) && ! defined (__MINGW32__) // 64 byte integer under Windows
 typedef unsigned __int64   U64;
 typedef __int64            I64;
-#else                                          // 64 byte integer elsewhere ... 
-typedef unsigned long long U64;
-typedef long long          I64;
+#else                                          // 64 byte integer elsewhere ...
+typedef uint64_t           U64;
+typedef int64_t            I64;
 #endif
 
 typedef float              F32;
@@ -94,9 +96,9 @@ typedef union U64I64F64 { U64 u64; I64 i64; F64 f64; } U64I64F64;
 #define U32_MIN            ((U32)0x0)            // 0
 #define U32_MAX            ((U32)0xFFFFFFFF)     // 4294967295
 #define U32_MAX_MINUS_ONE  ((U32)0xFFFFFFFE)     // 4294967294
-#if defined(WIN32)            // 64 byte unsigned int constant under Windows 
+#if defined(WIN32)            // 64 byte unsigned int constant under Windows
 #define U32_MAX_PLUS_ONE   0x0000000100000000    // 4294967296
-#else                         // 64 byte unsigned int constant elsewhere ... 
+#else                         // 64 byte unsigned int constant elsewhere ...
 #define U32_MAX_PLUS_ONE   0x0000000100000000ull // 4294967296
 #endif
